@@ -1,4 +1,4 @@
 #!/usr/bin/env python
 from .basemodule import *
 __all__ = []
-__all__ += ['ConvenientSpecMixin', 'SpecIOMixin', 'SpecSDSS', 'SpecIRAF', 'SpecLAMOST', 'SpecEuclid1d']
+__all__ += ['ConvenientSpecMixin', 'SpecIOMixin', 'SpecPandasRow', 'SpecSparcl', 'SpecSDSS', 'SpecIRAF', 'SpecLAMOST', 'SpecEuclid1d']
